make a scenario class
