+++ message_ix/__init__.py
@@ -1,8 +1,10 @@
 import os
 import re
 
-from message_ix import default_paths
 import ixmp.model_settings as model_settings
+
+from message_ix import default_paths
+from message_ix.core import *
 
 
 model_file = os.path.join(default_paths.model_path(), '{model}_run.gms')
