+++ message_ix/utils.py
@@ -31,78 +31,6 @@ def isscalar(x):
     return not isinstance(x, collections.Iterable) or isstr(x)
 
 
-def add_spatial_sets(scenario, data):
-    """Add sets related to spatial dimensions of the model
-
-    Parameters
-    ----------
-    scenario : ixmp.Scenario
-    data : dict or other
-
-    Examples
-    --------
-    data = {'country': 'Austria'}
-    data = {'country': ['Austria', 'Germany']}
-    data = {'country': {'Austria': {'state': ['Vienna', 'Lower Austria']}}}
-    """
-    nodes = []
-    levels = []
-    hierarchy = []
-
-    def recurse(k, v, parent='World'):
-        if isinstance(v, collections.Mapping):
-            for _parent, _data in v.items():
-                for _k, _v in _data.items():
-                    recurse(_k, _v, parent=_parent)
-
-        level = k
-        children = [v] if isscalar(v) else v
-        for child in children:
-            hierarchy.append([level, child, parent])
-            nodes.append(child)
-        levels.append(level)
-
-    for k, v in data.items():
-        recurse(k, v)
-
-    scenario.add_set("node", nodes)
-    scenario.add_set("lvl_spatial", levels)
-    scenario.add_set("map_spatial_hierarchy", hierarchy)
-
-
-def add_temporal_sets(scenario, data):
-    """Add sets related to temporal dimensions of the model
-
-    Parameters
-    ----------
-    scenario : ixmp.Scenario
-    data : dict or other
-
-    Examples
-    --------
-    data = {'year': [2010, 2020]}
-    data = {'year': [2010, 2020], 'firstmodelyear': 2020}
-    """
-    if 'year' not in data:
-        raise ValueError('"year" must be in temporal sets')
-    horizon = data['year']
-    scenario.add_set("year", horizon)
-
-    first = data['firstmodelyear'] if 'firstmodelyear' in data else horizon[0]
-    scenario.add_set("cat_year", ["firstmodelyear", first])
-
-
-def vintage_and_active_years(scenario):
-    """Return a 2-tuple of valid pairs of vintage years and active years for
-    use with data input.
-    """
-    horizon = scenario.set('year')
-    combinations = itertools.product(horizon, horizon)
-    year_pairs = [(y_v, y_a) for y_v, y_a in combinations if y_v <= y_a]
-    v_years, a_years = zip(*year_pairs)
-    return v_years, a_years
-
-
 def make_df(base, **kwargs):
     """Combine existing data with a series of new data defined in kwargs.
